Switch to a session factory instead
